Update patch set 1
Patch Set 1:

> Uploaded patch set 1.

This goes probably beyond what a client library should do, but our goal is rather to provide programmatic access to KorAP, for which this feature is very useful.
It might be a compromise to offer this in another package, but it's hardly worth it, because on top op this, I would only add the links from data points to corresponding KorAP requests (as in the demos).

Patch-set: 1